requirements.txt
# Core Dependencies
numpy==2.3.1
pandas==2.3.1
pydantic==2.11.7
redis==6.2.0
PyYAML==6.0.2

# Machine Learning
torch==2.7.1
transformers==4.53.1
scikit-learn
lightgbm==4.6.0
huggingface-hub>=0.33.2

# API and Web
fastapi==0.116.0
uvicorn==0.35.0
httpx==0.28.1

# Database
sqlalchemy==1.4.54
alembic==1.16.3

# Time Series
TA-Lib
yfinance
tushare==1.4.21

# Real-time Data
websockets==11.0.3
aiokafka==0.12.0
orjson==3.10.15

# Crypto Data
ccxt
requests==2.32.4
beautifulsoup4==4.13.4

# Sentiment Analysis
vaderSentiment==3.3.2

# Monitoring
prometheus_client==0.22.1
structlog==25.4.0

# Development
pytest==8.4.1
pytest-asyncio
pytest-benchmark
black
flake8
isort

# Configuration Management
python-dotenv==1.1.1
click==8.2.1

# Networking
aiohttp==3.12.13

# Financial Data
python-dateutil==2.9.0.post0

# Hot-reload file watching
watchdog>=3.0.0

# Workflow automation 
airflow==3.0.2

# Model deployment
onnx==1.18.0
onnxruntime==1.22.0

# Statistical analysis
scipy==1.16.0

# Distributed systems
celery

# Task scheduling
croniter==6.0.0

# Data validation
marshmallow

# Async utilities
asyncio-mqtt
aiofiles

# Caching
joblib==1.5.1

# Time utilities
pendulum==3.1.0

# Configuration
typer==0.16.0

# Utilities
tqdm==4.67.1
tabulate==0.9.0

# Cloud storage
boto3==1.39.3

# Security
cryptography==45.0.5

# Logging
colorlog==6.9.0

# Testing utilities
freezegun
responses

# Documentation
sphinx
sphinx-rtd-theme

# Code quality
mypy
coverage
pytest-cov

# Performance profiling
psutil==7.0.0
memory-profiler

# New Dependencies for Enhanced Features
# Async Redis for high-performance operations
aioredis>=2.0.0

# ONNX Runtime for fast ML inference (already present but ensure latest)
# onnxruntime==1.22.0

# NautilusTrader for high-performance backtesting and execution
nautilus-trader>=1.194.0
pyarrow>=15.0.0

# Advanced numerical computing for GLS and statistical analysis
# scipy==1.16.0 (already present)

# Enhanced data validation
pydantic>=2.0.0

# Machine Learning - Additional libraries
river>=0.21.0  # Online machine learning
contextualbandits>=0.19.0  # Multi-armed bandit algorithms
binance-connector-python>=3.9.0  # Binance official connector
peft>=0.8.0  # Parameter Efficient Fine-Tuning for LoRA

# Clean up duplicate dependencies
# PyYAML>=6.0.0  # Already specified above as PyYAML==6.0.2
# redis>=4.0.0   # Already specified above as redis==6.2.0
# psutil>=5.9.0  # Already specified above as psutil==7.0.0
# pytest>=7.0.0 # Already specified above as pytest==8.4.1 
